https://saugamedev.itch.io/hover-4-hire
In Hover 4 Hire, the player must drive a Hover Taxi around a futuristic island city, with each second spent driving depleting the taxi's internal battery.  If the battery runs out, the game ends. Each waypoint, however, has a built-in remote charging station in the road, so every pickup or drop-off means a boost of power.  The further apart the waypoints get, however, the more likely it is that the driver will fail to make it in time before running out of juice. Each pickup, drop-off, and second spent driving with a passenger earns the driver money.  Money is the score of the game; high scores mean higher income. The money is purely for score; there are no power-ups or upgrades to buy.