https://saugamedev.itch.io/glass-cannon
Glass Cannon is a 2D shooter with the top-down perspective. The game takes place in an arena wearing the mask of an operating theatre with a crowd watching. The game has two distinct stages; in the first stage, you have a straight-jacket on, and taking hits from enemies boosts one of three stats: health, speed, and power. In the second stage, the straight-jacket comes off and it's time to deliver the pain! With your trusty sidearm, you will attempt to last against the hordes of the evil nurses and doctors. Perform well, and the crowd may give you a gift. The goal is to survive as long as possible.